Global funciton added to handle creating hashed course members
